Refactor business logic from ProfileSnapshot + removed unused files
